Create test file for dss/storage/bundles.py
