+++ tests/infra/mock_storage_handler.py
@@ -1,0 +1,86 @@
+class MockStorageHandler:
+    bundle_list = ['bundles/00000000-1112-416a-bbcd-8a261c10b121.2019-02-26T035448.489896Z',
+                   'bundles/00000000-1112-4858-bf17-513dc2d05863.2019-02-26T033907.789762Z',
+                   'bundles/00000000-1112-4858-bf17-513dc2d05863.2019-05-23T220340.829000Z',
+                   'bundles/00000000-1112-4858-bf17-513dc2d05863.dead',
+                   'bundles/00000000-1112-5145-9560-430930445071.2018-10-09T002318.558884Z',
+                   'bundles/00000000-1112-525e-95d1-12b886550da9.2019-02-26T033907.789762Z',
+                   'bundles/00000000-1112-56b3-a467-0b0ac8208c6b.2019-02-26T033414.851898Z',
+                   'bundles/00000000-1112-56b3-a467-0b0ac8208c6b.2019-02-26T033414.851898Z.dead',
+                   'bundles/00000000-1112-5719-a070-1fcb25a57519.2018-10-18T203641.653105Z',
+                   'bundles/00000000-1112-5719-a070-1fcb25a57519.2018-10-18T203641.653105Z.dead',
+                   'bundles/00000000-1112-6146-8623-96cd047ca3df.2019-02-26T033558.843528Z',
+                   'bundles/00000000-1112-6265-8bf4-bb7bd2b38313.2019-02-26T035253.735430Z',
+                   'bundles/00000000-1112-6372-b236-21e6359cb933.2019-02-26T035613.863196Z',
+                   'bundles/00000000-1112-6452-94d9-293fcb97ba32.2019-05-10T110312.514000Z',
+                   'bundles/00000000-1112-6452-94d9-293fcb97ba32.dead',
+                   'bundles/00000000-1112-7197-967c-d6a33233f65e.2018-09-13T131642.918470Z',
+                   'bundles/00000000-1112-7670-9dad-92c9fb6534bb.2019-02-26T033443.407780Z',
+                   'bundles/00000000-1112-7ee4-aefd-6eb2c8cad783.2018-10-08T201847.591231Z',
+                   'bundles/00000000-1112-7ee4-aefd-6eb2c8cad783.2018-10-08T201847.591231Z.dead',
+                   'bundles/00000000-1112-7ee4-aefd-6eb2c8cad786.2018-10-08T201847.591231Z',
+                   'bundles/00000000-1113-416a-bbcd-8a261c10b121.2019-02-26T035448.489896Z',
+                   'bundles/00000000-1113-416a-bbcd-8a261c10b121.2019-02-26T035448.489896Z.dead',
+                   'bundles/00000000-1113-4858-bf17-513dc2d05863.2019-02-26T033907.789762Z',
+                   'bundles/00000000-1113-4858-bf17-513dc2d05863.2019-05-23T220340.829000Z',
+                   'bundles/00000000-1113-4858-bf17-513dc2d05863.dead',
+                   'bundles/00000000-1113-5145-9560-430930445071.2018-10-09T002318.558884Z',
+                   'bundles/00000000-1113-525e-95d1-12b886550da9.2019-02-26T033907.789762Z',
+                   'bundles/00000000-1113-56b3-a467-0b0ac8208c6b.2019-02-26T033414.851898Z',
+                   'bundles/00000000-1113-56b3-a467-0b0ac8208c6b.2019-02-26T033414.851898Z.dead',
+                   'bundles/00000000-1113-5719-a070-1fcb25a57519.2018-10-18T203641.653105Z',
+                   'bundles/00000000-1113-5719-a070-1fcb25a57519.2018-10-18T203641.653105Z.dead',
+                   'bundles/00000000-1113-6146-8623-96cd047ca3df.2019-02-26T033558.843528Z',
+                   'bundles/00000000-1113-6265-8bf4-bb7bd2b38313.2019-02-26T035253.735430Z',
+                   'bundles/00000000-1113-6372-b236-21e6359cb933.2019-02-26T035613.863196Z',
+                   'bundles/00000000-1113-6452-94d9-293fcb97ba32.2019-05-10T110312.514000Z',
+                   'bundles/00000000-1113-6452-94d9-293fcb97ba32.dead',
+                   'bundles/00000000-1113-7197-967c-d6a33233f65e.2018-09-13T131642.918470Z',
+                   'bundles/00000000-1113-7670-9dad-92c9fb6534bb.2019-02-26T033443.407780Z',
+                   'bundles/00000000-1113-7ee4-aefd-6eb2c8cad783.2018-10-08T201847.591231Z',
+                   'bundles/00000000-1113-7ee4-aefd-6eb2c8cad786.2018-10-08T201847.591231Z']
+
+    dead_bundles = ['bundles/00000000-1112-4858-bf17-513dc2d05863.2019-02-26T033907.789762Z',
+                    'bundles/00000000-1112-4858-bf17-513dc2d05863.2019-05-23T220340.829000Z',
+                    'bundles/00000000-1112-4858-bf17-513dc2d05863.dead',
+                    'bundles/00000000-1112-56b3-a467-0b0ac8208c6b.2019-02-26T033414.851898Z',
+                    'bundles/00000000-1112-56b3-a467-0b0ac8208c6b.2019-02-26T033414.851898Z.dead',
+                    'bundles/00000000-1112-5719-a070-1fcb25a57519.2018-10-18T203641.653105Z',
+                    'bundles/00000000-1112-5719-a070-1fcb25a57519.2018-10-18T203641.653105Z.dead',
+                    'bundles/00000000-1112-6452-94d9-293fcb97ba32.2019-05-10T110312.514000Z',
+                    'bundles/00000000-1112-6452-94d9-293fcb97ba32.dead',
+                    'bundles/00000000-1112-6e52-94d9-293fcb97ba32.2019-05-10T110312.514000Z',
+                    'bundles/00000000-1112-6e52-94d9-293fcb97ba32.dead',
+                    'bundles/00000000-1112-7ee4-aefd-6eb2c8cad783.2018-10-08T201847.591231Z',
+                    'bundles/00000000-1112-7ee4-aefd-6eb2c8cad783.2018-10-08T201847.591231Z.dead',
+                    'bundles/00000000-1113-416a-bbcd-8a261c10b121.2019-02-26T035448.489896Z',
+                    'bundles/00000000-1113-416a-bbcd-8a261c10b121.2019-02-26T035448.489896Z.dead',
+                    'bundles/00000000-1113-4858-bf17-513dc2d05863.2019-02-26T033907.789762Z',
+                    'bundles/00000000-1113-4858-bf17-513dc2d05863.2019-05-23T220340.829000Z',
+                    'bundles/00000000-1113-4858-bf17-513dc2d05863.dead',
+                    'bundles/00000000-1113-56b3-a467-0b0ac8208c6b.2019-02-26T033414.851898Z',
+                    'bundles/00000000-1113-56b3-a467-0b0ac8208c6b.2019-02-26T033414.851898Z.dead',
+                    'bundles/00000000-1113-5719-a070-1fcb25a57519.2018-10-18T203641.653105Z',
+                    'bundles/00000000-1113-5719-a070-1fcb25a57519.2018-10-18T203641.653105Z.dead',
+                    'bundles/00000000-1113-6452-94d9-293fcb97ba32.2019-05-10T110312.514000Z',
+                    'bundles/00000000-1113-6452-94d9-293fcb97ba32.dead',
+                    ]
+
+    test_per_page = [{"size": 10,
+                      "last_good_bundle": {"uuid": "00000000-1113-5145-9560-430930445071",
+                                           "version": "2018-10-09T002318.558884Z"}},
+                     {"size": 12,
+                      "last_good_bundle": {"uuid": "00000000-1113-6146-8623-96cd047ca3df",
+                                           "version": "2019-02-26T033558.843528Z"}},
+                     {"size": 14,
+                      "last_good_bundle": {"uuid": "00000000-1113-6372-b236-21e6359cb933",
+                                           "version": "2019-02-26T035613.863196Z"}}]
+
+    def list_v2(self, *args, **kwargs):
+        search_after = kwargs.get('start_after_key')
+        bundle_list = self.bundle_list
+        if search_after:
+            idx = self.bundle_list.index(search_after)
+            bundle_list = self.bundle_list[idx + 1:-1]
+        list_tuples = [(x, None) for x in bundle_list]
+        return iter(list_tuples)
