+++ tests/test_bundle.py
@@ -33,9 +33,9 @@ from dss.storage.blobstore import test_object_exists
 from dss.storage.hcablobstore import compose_blob_key
 from dss.util.version import datetime_to_version_format
 from dss.storage.bundles import get_bundle_manifest
-from tests.infra import DSSAssertMixin, DSSUploadMixin, ExpectedErrorFields, get_env, testmode, TestAuthMixin
+from tests.infra import (DSSAssertMixin, DSSUploadMixin, ExpectedErrorFields, get_env, testmode, TestAuthMixin,
+                         MockStorageHandler)
 from tests.infra.server import ThreadedLocalServer, MockFusilladeHandler
-from tests.test_utils import MockStorageHandler
 from tests import eventually, get_auth_header
